+++ tests/__fixtures__/projects/libs/greeter/src/greeter/core.clj
@@ -1,0 +1,4 @@
+(ns greeter.core)
+
+(defn greet [name]
+  (str "Hello, " (or name "world") "!"))