+++ tests/__fixtures__/projects/libs/testkit/src/testkit/helpers.clj
@@ -1,0 +1,3 @@
+(ns testkit.helpers)
+
+(defn fixture [x] x)